client/requirements.txt
docker-py==1.10.6
numpy==1.19.5
psutil==5.9.0
